+++ library/Kathu/World/Field.hs
@@ -1,6 +1,7 @@
-{-# LANGUAGE BangPatterns #-}
-{-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE BangPatterns  #-}
+{-# LANGUAGE LambdaCase    #-}
 {-# LANGUAGE TupleSections #-}
+{-# LANGUAGE UnboxedTuples #-}
 
 module Kathu.World.Field where
 
@@ -13,7 +14,7 @@ import qualified Data.Map.Strict as Map
 import Data.Maybe (maybe)
 import qualified Data.Vector as Vec
 import qualified Data.Vector.Unboxed.Mutable as UMVec
-import Linear.V3 (V3(..))
+import Linear.V2 (V2(..))
 
 import Kathu.Util.MultiDimVector
 import Kathu.World.Tile
@@ -22,80 +23,116 @@ unitsPerTile :: Num a => a
 unitsPerTile = 16
 
 -- Fields are 32 by 32 tiles
-fieldSize :: Num a => a
-fieldSize = 32
+fieldDim :: Num a => a
+fieldDim = 32
 
--- Fields have a height of 4 tiles tall
-fieldHeight :: Num a => a
-fieldHeight = 4
+type FieldSet = Map (V2 Int) Field
 
-type FieldSet = Map (V3 Int) Field
+data Layer = Foreground | Background deriving (Show, Eq)
 
-newtype Field = Field
-    { fieldData :: MVector RealWorld (TileState)
+data Field = Field
+    { fieldTileForeground :: {-# UNPACK #-} !(MVector RealWorld (TileState))
+    , fieldTileBackground :: {-# UNPACK #-} !(MVector RealWorld (TileState))
     --, fieldIndRenders :: (IOVector Bool)
     --, fieldPreRenders :: [Image]
     }
 
 mkField :: MonadIO m => m (Field)
-mkField = liftIO . fmap Field . UMVec.replicate size . mkTileState $ emptyTile
-    where size = fieldSize * fieldSize * fieldHeight
+mkField = liftIO $ Field <$> UMVec.replicate size emptyTS <*> UMVec.replicate size emptyTS
+    where size    = fieldDim * fieldDim
+          emptyTS = mkTileState emptyTile
 
-indexFromCoord :: V3 Int -> Int
-indexFromCoord (V3 x y z) = z * (fieldSize * fieldSize) + x * fieldSize + y
+--------------------------
+-- Int -> Int Functions --
+--------------------------
 
-fieldContainingCoord :: RealFrac a => V3 a -> V3 Int
-fieldContainingCoord (V3 x y z) = floor <$> (V3 (x / (unitsPerTile * fieldSize)) (y / (unitsPerTile * fieldSize)) (z / (unitsPerTile * fieldHeight)))
+{-# INLINE indexFromCoord #-}
+indexFromCoord :: Int -> Int -> Int
+indexFromCoord !x !y = x * fieldDim + y
+
+fieldContainingCoord :: RealFrac a => a -> a -> (# Int, Int #)
+fieldContainingCoord !x !y = (# getCoord x, getCoord y #)
+    where getCoord = floor . (/(unitsPerTile * fieldDim))
+
+fetchTileState :: MonadIO m => Layer -> Int -> Int -> Field -> m (TileState)
+fetchTileState Foreground !x !y (Field fgTiles _) = liftIO $ UMVec.read fgTiles (indexFromCoord x y)
+fetchTileState Background !x !y (Field _ bgTiles) = liftIO $ UMVec.read bgTiles (indexFromCoord x y)
+
+setTileState :: MonadIO m => Layer -> Int -> Int -> TileState -> Field -> m ()
+setTileState Foreground !x !y t (Field fgTiles _) = liftIO $ UMVec.write fgTiles (indexFromCoord x y) t
+setTileState Background !x !y t (Field _ bgTiles) = liftIO $ UMVec.write bgTiles (indexFromCoord x y) t
+
+----------------------
+-- V2 Int Functions --
+----------------------
+
+{-# INLINE indexFromCoordV2 #-}
+indexFromCoordV2 :: V2 Int -> Int
+indexFromCoordV2 (V2 !x !y) = x * fieldDim + y
+
+fieldContainingCoordV2 :: RealFrac a => V2 a -> (# Int, Int #)
+fieldContainingCoordV2 (V2 !x !y) = (# getCoord x, getCoord y #)
+    where getCoord = floor . (/(unitsPerTile * fieldDim))
+
+fetchTileStateV2 :: MonadIO m => Layer -> V2 Int -> Field -> m (TileState)
+fetchTileStateV2 Foreground (V2 !x !y) (Field fgTiles _) = liftIO $ UMVec.read fgTiles (indexFromCoord x y)
+fetchTileStateV2 Background (V2 !x !y) (Field _ bgTiles) = liftIO $ UMVec.read bgTiles (indexFromCoord x y)
+
+setTileStateV2 :: MonadIO m => Layer -> V2 Int -> TileState -> Field -> m ()
+setTileStateV2 Foreground (V2 !x !y) t (Field fgTiles _) = liftIO $ UMVec.write fgTiles (indexFromCoord x y) t
+setTileStateV2 Background (V2 !x !y) t (Field _ bgTiles) = liftIO $ UMVec.write bgTiles (indexFromCoord x y) t
 
 -- Field Coord -> Local Coord in Field -> World Coord
-worldCoordFromTile :: Num a => V3 Int -> V3 Int -> V3 a
-worldCoordFromTile (V3 fx fy fz) (V3 lx ly lz) = V3 (conv fieldSize fx lx) (conv fieldSize fy ly) (conv fieldHeight fz lz)
-    where conv s f l = unitsPerTile * fromIntegral ((f * s) + l)
+worldCoordFromTileCoord :: Num a => V2 Int -> V2 Int -> V2 a
+worldCoordFromTileCoord (V2 !fx !fy) (V2 !lx !ly) = V2 (conv fieldDim fx lx) (conv fieldDim fy ly)
+    where conv !s !f !l = unitsPerTile * fromIntegral ((f * s) + l)
 
-fetchTileState :: MonadIO m => V3 Int -> Field -> m (TileState)
-fetchTileState v (Field tilest) = liftIO $ UMVec.read tilest (indexFromCoord v)
-
-setTileState :: MonadIO m => V3 Int -> TileState -> Field -> m ()
-setTileState v t (Field tilest) = liftIO $ UMVec.write tilest (indexFromCoord v) t
+----------
+-- Misc --
+----------
 
 foreachTile :: MonadIO m => (TileState -> m a) -> Field -> m ()
-foreachTile f (Field tilest) = go 0
-    where len = UMVec.length tilest
-          go i | i == len  = pure ()
-               | otherwise = liftIO (UMVec.read tilest i) >>= f >> go (i + 1)
+foreachTile f (Field fgTiles bgTiles) = go bgTiles 0 >> go fgTiles 0
+    where len = UMVec.length fgTiles
+          go tiles !i | i == len  = pure ()
+                      | otherwise = liftIO (UMVec.unsafeRead tiles i) >>= f >> go tiles (i + 1)
 
-fieldFoldM :: MonadIO m => (a -> V3 Int -> TileState -> m a) -> a -> Field -> m a
-fieldFoldM f !acc (Field tiles) = go 0 0 0 acc
-    where go !x !y !z !b | z == fieldHeight = pure b
-                         | y == fieldSize   = go 0 0 (z + 1) b
-                         | x == fieldSize   = go 0 (y + 1) z b
-                         | otherwise        = let v = V3 x y z in liftIO (UMVec.read tiles (indexFromCoord v)) >>= f b v >>= go (x + 1) y z
+-- | Folds through all present tiles in the field monadically, with position and layer information; skips empty tiles
+fieldFoldM :: MonadIO m => (a -> Layer -> V2 Int -> TileState -> m a) -> a -> Field -> m a
+fieldFoldM f !acc (Field fgTiles bgTiles) = go Background bgTiles 0 0 acc >>= go Foreground fgTiles 0 0
+    where go !layer tiles !x !y !b
+              | y == fieldDim            = pure b
+              | x == fieldDim            = go layer tiles 0 (y + 1) b
+              | otherwise                = liftIO (UMVec.unsafeRead tiles $ indexFromCoord x y) >>= ignoreEmpty b (f b layer (V2 x y)) >>= go layer tiles (x + 1) y
+          ignoreEmpty b action !t        = if t^.tile == emptyTileID then pure b else action t
 
--- Conversion
+----------------
+-- Conversion --
+----------------
 
-mkFields :: MonadIO m => Int -> Int -> Int -> m (Vector3D (Field))
-mkFields x y z = repliVecM x (repliVecM y (repliVecM z mkField))
+mkFields :: MonadIO m => Int -> Int -> m (Vector2D (Field))
+mkFields x y = repliVecM x (repliVecM y mkField)
     where repliVecM n = fmap Vec.fromList . replicateM n
    
-fromTileVector3D :: MonadIO m => Vector3D (Tile g) -> m FieldSet
-fromTileVector3D tiles = if zLayers == 0 || yLayers == 0 || xLayers == 0 then empty else buildNew 
-    where empty                       = Map.singleton (V3 0 0 0) <$> mkField
+fromTileVector2D :: MonadIO m => Vector2D (Tile g) -> Vector2D (Tile g) -> m FieldSet
+fromTileVector2D fgTiles bgTiles = if yLayers == 0 || xLayers == 0 then empty else buildNew 
+    where empty                       = Map.singleton (V2 0 0) <$> mkField
           minFields :: (Integral a) => Float -> a -> Int
           minFields s                 = ceiling . (/s) . fromIntegral
-          getContField x y z = fieldContainingCoord $ (fromIntegral :: Int -> Float) <$> (V3 x y z)
-          -- tiles are stored as z x y, rather than x y z
-          (zLen, yLen, xLen)          = (Vec.length tiles, Vec.length (tiles Vec.! 0), Vec.length ((tiles Vec.! 0) Vec.! 0))
-          (zLayers, yLayers, xLayers) = (minFields fieldHeight zLen, minFields fieldSize yLen, minFields fieldSize xLen)
-          buildNew = liftIO (new3DWith xLayers yLayers zLayers Nothing) >>= go 0 0 0 >>= convertToMap
-          go x y z v | z == zLen = pure v
-                     | y == yLen = go 0 0 (z + 1) v
-                     | x == xLen = go 0 (y + 1) z v
-                     | otherwise = (runForTile (read3D z y x tiles) writeTile) >> go (x + 1) y z v
-              where (V3 fx fy fz) = getContField x y z
-                    runForTile t f | t^.tileID == emptyTileID = pure () -- do nothing if wanting to write empty tile
-                                   | otherwise                = liftIO (mRead3D fx fy fz v) >>= \case
+          getContField x y   = fieldContainingCoord ((fromIntegral :: Int -> Double) x) (fromIntegral y)
+          -- tiles are stored as x y, rather than x y
+          (yLen, xLen)       = (Vec.length fgTiles, Vec.length (fgTiles Vec.! 0))
+          (yLayers, xLayers) = (minFields fieldDim yLen, minFields fieldDim xLen)
+          buildNew = liftIO (new2DWith xLayers yLayers Nothing) >>= go Foreground fgTiles 0 0 >>= go Background bgTiles 0 0 >>= convertToMap
+          go :: MonadIO m => Layer -> Vector2D (Tile g) -> Int -> Int -> IOVector2D (Maybe Field) -> m (IOVector2D (Maybe Field))
+          go !layer !tiles !x !y v | y == yLen = pure v
+                                   | x == xLen = go layer tiles 0 (y + 1) v
+                                   | otherwise = (runForTile (read2D y x tiles) writeTile) >> go layer tiles (x + 1) y v
+              where (# fx, fy #) = getContField x y
+                    runForTile t f | t^.tileID == emptyTileID = pure () -- do nothing if wanting to write empty tile, since Fields are initialized with all emptyTile
+                                   | otherwise                = liftIO (mRead2D fx fy v) >>= \case
                                        Just (m) -> f t m
-                                       Nothing  -> mkField >>= \field -> liftIO (mWrite3D fx fy fz (Just field) v) >> f t field
-                    writeTile t field = mkTileStateWithMetadata t >>= \newTileState -> setTileState (V3 x y z) newTileState field
-          convertToMap iovec = fmap Map.fromList . liftIO .  miFoldl3D foldf [] $ iovec
-              where foldf acc x y z cur = pure $ maybe acc ((:acc) . (V3 x y z,)) cur
+                                       Nothing  -> mkField >>= \field -> liftIO (mWrite2D fx fy (Just field) v) >> f t field
+                    writeTile t field = mkTileStateWithMetadata t >>= \newTileState -> setTileState layer x y newTileState field
+          convertToMap iovec = fmap Map.fromList . liftIO .  miFoldl2D foldf [] $ iovec
+              where foldf acc x y cur = pure $ maybe acc ((:acc) . (V2 x y,)) cur