+++ library/Kathu/World/WorldSpace.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE BangPatterns         #-}
 {-# LANGUAGE DataKinds            #-}
+{-# LANGUAGE DeriveGeneric        #-}
 {-# LANGUAGE ExplicitForAll       #-}
 {-# LANGUAGE FlexibleContexts     #-}
 {-# LANGUAGE FlexibleInstances    #-}
@@ -8,6 +9,7 @@
 {-# LANGUAGE ScopedTypeVariables  #-}
 {-# LANGUAGE TupleSections        #-}
 {-# LANGUAGE TypeOperators        #-}
+{-# LANGUAGE UnboxedTuples        #-}
 {-# LANGUAGE UndecidableInstances #-}
 
 module Kathu.World.WorldSpace where
@@ -17,13 +19,15 @@ import Control.Monad (foldM)
 import Control.Monad.IO.Class (MonadIO)
 import Data.Aeson
 import Data.Aeson.Types (Parser, typeMismatch)
+import Data.Map (Map)
 import qualified Data.Map as Map
-import Data.Maybe (catMaybes, fromMaybe)
+import Data.Maybe (catMaybes, fromMaybe, maybe)
 import Data.Text (Text)
 import qualified Data.Text as T
 import Data.Vector (Vector)
 import qualified Data.Vector as Vec
-import Linear.V3 (V3(..))
+import GHC.Generics
+import Linear.V2 (V2(..))
 
 import Kathu.Entity.Components
 import Kathu.Entity.Item
@@ -34,34 +38,40 @@ import Kathu.Graphics.Palette
 import Kathu.Util.Apecs
 import Kathu.Util.Dependency
 import Kathu.Util.Flow ((>>>=))
-import Kathu.Util.MultiDimVector (fromList3D)
+import qualified Kathu.Util.MultiDimVector as MDVec
 import Kathu.Util.Types (Identifier, IDMap)
 import Kathu.World.Field
-import Kathu.World.Tile (Tile)
+import Kathu.World.Tile (Tile, emptyTile)
+
+data WorldVariable
+    = WorldBool Bool
+    | WorldDouble Double
+    | WorldInt Int
+      deriving (Show, Eq, Generic)
 
 data WorldSpace g = WorldSpace
-    { worldID       :: Identifier
-    , worldName     :: Text
-    , worldPalettes :: Vector Palette
-    , loadPoint     :: V3 Float
-    , worldEntities :: Vector (V3 Float, EntityPrototype g)
-    , worldItems    :: Vector (V3 Float, ItemStack g)
-    , worldFields   :: FieldSet
-    -- Some way to hold the fields that it possesses
+    { worldID        :: Identifier
+    , worldName      :: Text
+    , worldPalettes  :: Vector Palette
+    , loadPoint      :: V2 Float
+    , worldVariables :: IDMap WorldVariable
+    , worldEntities  :: Vector (V2 Float, EntityPrototype g)
+    , worldItems     :: Vector (V2 Float, ItemStack g)
+    , worldFields    :: FieldSet
     }
 
 emptyWorldSpace :: WorldSpace g
-emptyWorldSpace = WorldSpace "" "the void" (Vec.singleton (Palette black Nothing)) (V3 0 0 0) Vec.empty Vec.empty Map.empty
+emptyWorldSpace = WorldSpace "" "the void..." (Vec.singleton (Palette black Nothing)) (V2 0 0) Map.empty Vec.empty Vec.empty Map.empty
 
 -- right now we only consider horizontal fields; ones with different z depths are ignored
-fieldsSurrounding :: RealFrac a => V3 a -> WorldSpace g -> [(V3 Int, Field)]
+fieldsSurrounding :: RealFrac a => V2 a -> WorldSpace g -> [(V2 Int, Field)]
 fieldsSurrounding v ws = catMaybes $ readFields [] (ox - 1) (oy - 1)
     where fields    = worldFields ws
-          (V3 ox oy oz) = fieldContainingCoord v
+          (# ox, oy #) = fieldContainingCoordV2 v
           readFields !acc !x !y | y > oy + 1 = acc
                                 | x > ox + 1 = readFields acc 0 (y + 1)
                                 | otherwise  = readFields (((curV,) <$> Map.lookup curV fields):acc) (x + 1) y
-              where curV = V3 x y oz -- only consider same z level right now
+              where curV = V2 x y -- only consider same z level right now
 
 --------------------
 -- System Related --
@@ -76,13 +86,26 @@ initWorldSpace mkEntity ws = do
     mapM_ (\(pos, ety)  -> mkEntity ety >>= (flip set) (Position pos)) (worldEntities ws)
     mapM_ (\(pos, item) -> newEntityFromItem item pos) (worldItems ws)
 
-newEntityFromItem :: forall w m g. (MonadIO m, Get w m EntityCounter, HasEach w m '[Position, Render g]) => ItemStack g -> V3 Float -> SystemT w m Entity
+newEntityFromItem :: forall w m g. (MonadIO m, Get w m EntityCounter, HasEach w m '[Position, Render g]) => ItemStack g -> V2 Float -> SystemT w m Entity
 newEntityFromItem stack v = newEntity (Position v, itemIcon . stackItem $ stack)
           -- as of right now, count not considered; this will be added when picking up is implemented
 
 -------------------
 -- Serialization --
 -------------------
+
+instance ToJSON WorldVariable where
+    toJSON (WorldBool b)   = object ["type" .= ("bool" :: Text),   "value" .= b]
+    toJSON (WorldDouble d) = object ["type" .= ("double" :: Text), "value" .= d]
+    toJSON (WorldInt i)    = object ["type" .= ("int" :: Text),    "value" .= i]
+instance FromJSON WorldVariable where
+    parseJSON (Object v)   = v .: "type" >>= var
+        where var :: Text -> Parser WorldVariable
+              var "bool"   = WorldBool   <$> v .: "value"
+              var "double" = WorldDouble <$> v .: "value"
+              var "int"    = WorldInt    <$> v .: "value"
+              var e        = fail . concat $ ["Unable to parse world variable with type of ", show e]
+    parseJSON e            = typeMismatch "WorldVariable" e
 
 -- Instances and functions related to serializing the WorldSpace
 -- This is kept at the end due to the length and complexity of theses
@@ -93,12 +116,14 @@ instance ( s `CanProvide` (IDMap (EntityPrototype g))
          , MonadIO m
          ) => FromJSON (Dependency s m (WorldSpace g)) where
     parseJSON (Object v) = do
-        worldId  <- v .: "world-id"
-        wName    <- v .: "name"
-        palettes <- v .: "palettes"
-        loadPnt  <- (*unitsPerTile) <$> v .: "load-point"
-        layers :: [[[Char]]] <- v .: "layers"
-        legend <- do
+        worldId   <- v .: "world-id"
+        wName     <- v .: "name"
+        palettes  <- v .: "palettes"
+        loadPnt   <- (*unitsPerTile) <$> v .: "load-point"
+        variables <- v .:? "global-variables" .!= Map.empty
+        backgroundT :: Maybe [[Char]] <- v .:? "background-tiles"
+        foregroundT :: [[Char]]       <- v .: "foreground-tiles"
+        legend      :: Dependency s m (Map Char (Tile g)) <- do
             (keys, vals) :: ([Text], [Identifier]) <- ((unzip . Map.toList) <$> v .: "legend")
             let dLookup = dependencyMapLookupElseError :: String -> Identifier -> Dependency s m (Tile g)
              in pure . fmap (Map.fromList . zip (T.head <$> keys)) . flattenDependency . fmap (dLookup "Tile") $ vals
@@ -112,7 +137,15 @@ instance ( s `CanProvide` (IDMap (EntityPrototype g))
             parseEty = withObject "EntityPlacement" $ \obj -> (obj .: "entity" :: Parser Identifier) >>= pure . dependencyMapLookupElseError "Entity" 
         items    <- v .: "items" >>= parsePlacement parseJSON >>>= pure . Vec.fromList
         entities <- v .: "entities" >>= parsePlacement parseEty >>>= pure . Vec.fromList
-        let layersVec     = (\lgnd -> fromList3D . fmap (fmap (fmap (fromMaybe failIfNothing . (flip Map.lookup) lgnd))) $ layers) <$> legend
-            failIfNothing = error "Attempted to tile without a listing in the WorldSpace's legend"
-        pure $ WorldSpace worldId wName palettes loadPnt <$> entities <*> items <*> (liftDependency . fromTileVector3D =<< layersVec)
+        let failIfNothing = error "Attempted to tile without a listing in the WorldSpace's legend"
+            applyLegend tileList2D = (\lgnd -> MDVec.fromList2D . fmap (fmap (fromMaybe failIfNothing . (flip Map.lookup) lgnd)) $ tileList2D) <$> legend
+            fgVec :: Dependency s m (MDVec.Vector2D (Tile g))
+            fgVec = applyLegend foregroundT
+            -- if we weren't given a background map, we assume its the same dimensions as the foreground, but empty
+            bgVec :: Dependency s m (MDVec.Vector2D (Tile g))
+            bgVec = maybe ((((const emptyTile)<$>)<$>)<$> fgVec) applyLegend $ backgroundT
+            newFieldSet :: Dependency s m (FieldSet)
+            newFieldSet = do {fg <- applyLegend foregroundT; bg <- bgVec; liftDependency $ fromTileVector2D fg bg}
+
+        pure $ WorldSpace worldId wName palettes loadPnt variables <$> entities <*> items <*> newFieldSet
     parseJSON e          = typeMismatch "WorldSpace" e